renamed auth pages and routes
