+++ src/web/pages/auth/reset/new.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ResetPassword } from '../../components/resetpassword/ResetPassword';
+import { ResetPassword } from '../../../components/resetpassword/ResetPassword';
 // import { Layout } from '../../components/layout/index';
 // export const ForgotPasswordView: React.FC = () => {
 // export default function ForgotPasswordView() {
